move graph layout related types to GraphLayout.Types
